refactor: moved alot of code to seperate files
