+++ src/go.mod
@@ -4,7 +4,5 @@ go 1.15
 
 require (
 	github.com/fatih/color v1.10.0
-	github.com/hashicorp/go-version v1.2.1 // indirect
-	github.com/mitchellh/gox v1.0.1 // indirect
 	github.com/urfave/cli/v2 v2.3.0
 )
